requirements.txt
asgiref==3.8.1
Django==5.1
django-appconf==1.0.6
django-compressor==4.5.1
django-storages==1.14.4
djangorestframework==3.15.2
ffmpeg-python==0.2.0
future==1.0.0
pillow==10.4.0
rcssmin==1.1.2
rjsmin==1.2.2
sqlparse==0.5.1
